In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -4,21 +4,21 @@
  * main - Entry point
  *
  * Description: Prints the first 98 Fibonacci numbers, starting with 1 and 2
- *              separated by comma and space
+ * separated by comma and space
  *
  * Return: Always 0 (Success)
  */
 int main(void)
 {
-	unsigned int fib1 = 1, fib2 = 2, fib_next;
+	unsigned long fib1 = 1, fib2 = 2, fib_next;
 	int count;
 
-	printf("%u, %u", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		fib_next = fib1 + fib2;
-		printf(", %u", fib_next);
+		printf(", %lu", fib_next);
 
 		fib1 = fib2;
 		fib2 = fib_next;
